seperate context and provider
