+++ package.json
@@ -126,7 +126,6 @@
 		"merge-config": "2.0.0",
 		"mini-css-extract-plugin": "0.9.0",
 		"node-sass": "4.13.1",
-		"null-loader": "^3.0.0",
 		"postcss-loader": "3.0.0",
 		"prettier": "npm:wp-prettier@1.19.1",
 		"progress-bar-webpack-plugin": "1.12.1",
